some changes

+++ BotDaddy.py
@@ -506,8 +506,9 @@ async def clean_hang_bot_flood(m):
         hangbot_flood = hang_bot_flood[m.chat.id]
         for message in hangbot_flood:
             try:
-                await bot.delete_message(m.chat.id, message.message_id)
-                await bot.delete_message(m.chat.id, message.reply_to_message.message_id)
+                if len(message) == 1:
+                    await bot.delete_message(m.chat.id, message.message_id)
+                    await bot.delete_message(m.chat.id, message.reply_to_message.message_id)
             except (exceptions.MessageToDeleteNotFound, AttributeError):
                 continue
         try:
